ReadDir Override
Make wolfSSH_SFTPNAME_readdir defining SFTP_Name_readdir. It has to be
defined as a macro, but it can be defined to expand to its own name if
also a function with the same name is defined.

What matters is that it takes 3 arguments:

    1) the filesystem context as first argument;
    2) WDIR* as second argument;
    3) WS_SFTPNAME* as third argument

On successful execution, it returns WS_SUCCESS and the WS_SFTPNAME
structure pointed by the third argument will be filled with the
relevant info, otherwise a WS_* error code is returned.